tasks now handle --reextract themselves

+++ src/tasks/boost.cpp
@@ -55,7 +55,19 @@ void boost::do_clean(clean c)
 		{
 			if (is_set(c, clean::redownload))
 				run_tool(downloader(source_url(), downloader::clean));
+		}
 
+
+		if (is_set(c, clean::reextract))
+		{
+			cx().trace(context::reextract, "deleting {}", source_path());
+			op::delete_directory(cx(), source_path(), op::optional);
+			return;
+		}
+
+
+		if (!prebuilt())
+		{
 			if (is_set(c, clean::reconfigure))
 			{
 				op::delete_directory(cx(), source_path() / "bin.v2", op::optional);
