+++ src/tasks/pyqt.cpp
@@ -47,7 +47,17 @@ void pyqt::do_clean(clean c)
 		{
 			if (is_set(c, clean::redownload))
 				run_tool(downloader(source_url(), downloader::clean));
+		}
 
+		if (is_set(c, clean::reextract))
+		{
+			cx().trace(context::reextract, "deleting {}", source_path());
+			op::delete_directory(cx(), source_path(), op::optional);
+			return;
+		}
+
+		if (!prebuilt())
+		{
 			if (is_set(c, clean::rebuild))
 			{
 				op::delete_file(cx(),
